F! rvbuild: fix argument usage
This patch take the script back to original design. In the original
design, the script should be used like: `rvbuild --options args <pkg>`.

Due to the lack of the knowledge about how Perl handle options, I use a
workaround, which use option `--package/-p` to receive package name.

But later I discover that the GetOptions::Long module will take out the
argument after it parsed. So the rest part inside of the `@ARGV` array
must be package name.

Signed-off-by: Avimitin <avimitin@gmail.com>